PLAT-555: move happy-path tests in contract_calls_test.go
